requirements.txt
instagrapi>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
